dev(memory/drivers): refactor
