+++ Makefile
@@ -11,7 +11,8 @@ CFLAGS+=-Wall -Wextra # -Werror
 CFLAGS+=-I./include -I./lib/libft
 
 SRC_INCLUDE_DIRS := $(shell find src -type d)
-CFLAGS += $(addprefix -I, $(SRC_INCLUDE_DIRS))
+INCLUDE_DIRS := $(shell find include -type d)
+CFLAGS += $(addprefix -I, $(INCLUDE_DIRS)) $(addprefix -I, $(SRC_INCLUDE_DIRS))
 
 CXX=i686-linux-gnu-g++
 CXXFLAGS=-ffreestanding -fno-builtin -fno-exceptions -fno-stack-protector -fno-rtti
@@ -58,7 +59,7 @@ all: $(BUILDDIR)/boot.iso
 else
 all:
 	docker run --rm -t -v .:/kfs -e IN_DOCKER=1 $(DOCKERIMAGENAME):$(DOCKERIMAGETAG)
-	
+
 endif
 
 $(BUILDDIR)/boot.iso: $(ISODIR)/boot/kernel $(ISODIR)/boot/grub/grub.cfg
