+++ include/memory/boot_allocator.h
@@ -1,16 +1,19 @@
 #pragma once
 
 // ============================================================================
-// IMCLUDES
+// INCLUDES
 // ============================================================================
 
-#include <memory.h>
+#include "mb2_info.h"
+#include <types.h>
 
 // ============================================================================
 // DEFINE AND MACRO
 // ============================================================================
 
 // Defines
+
+enum allocator_state { ACTIVE = 0, FROZEN };
 
 // Macros
 
@@ -20,24 +23,34 @@
 
 // Enums
 
+enum mem_type { FREE_MEMORY = 0, RESERVED_MEMORY, HOLES_MEMORY };
+
 // Structures
+
+struct region_s {
+	uintptr_t start;
+	uintptr_t end;
+};
+
+struct boot_allocator;
 
 // Typedefs
 
-typedef struct page page_t;
-typedef void (*pages_foreach_fn)(page_t *page, void *data);
+typedef struct region_s       region_t;
+typedef struct boot_allocator boot_allocator_t;
 
 // ============================================================================
 // VARIABLES GLOBALES
 // ============================================================================
 
 // ============================================================================
-// INTERNAL APIs
+// EXTERNAL APIs
 // ============================================================================
 
-static inline page_t *last_page(void);
-static inline page_t *first_page(void);
-static uint32_t       page_get_appropriate_flag(uintptr_t addr_start);
-static void           count_free_pages(page_t *page, void *counter);
-static void           count_reserved_pages(page_t *page, void *counter);
-static void           page_descriptor_foreach(pages_foreach_fn handler, void *data);
+bool      boot_allocator_range_overlaps(uintptr_t start, uintptr_t end, enum mem_type type);
+void     *boot_alloc(uint32_t size);
+void      boot_allocator_freeze(void);
+void      boot_allocator_printer(void);
+void      boot_allocator_init(multiboot_tag_mmap_t *mmap, uint8_t *mmap_end);
+uint32_t  boot_allocator_get_region_count(enum mem_type type);
+region_t *boot_allocator_get_region(enum mem_type type);
