+++ include/memory/memory.h
@@ -1,0 +1,42 @@
+#pragma once
+
+// ============================================================================
+// INCLUDES
+// ============================================================================
+
+#include "boot_allocator.h"
+#include "buddy.h"
+#include "page.h"
+
+// ============================================================================
+// DEFINE AND MACRO
+// ============================================================================
+
+// Defines
+
+// Macros
+
+// ============================================================================
+// STRUCT
+// ============================================================================
+
+// Enums
+
+// Structures
+
+// Typedefs
+
+// ============================================================================
+// VARIABLES GLOBALES
+// ============================================================================
+
+// (aucune variable globale à exposer ici)
+
+// ============================================================================
+// EXTERNAL APIs
+// ============================================================================
+
+// API mémoire globale (à adapter selon ton implémentation)
+// void *kmalloc(size_t size);
+// void  kfree(void *ptr);
+// void *krealloc(void *ptr, size_t new_size);
