+++ src/memory/internal/boot_allocator.h
@@ -4,7 +4,7 @@
 // IMCLUDES
 // ============================================================================
 
-#include <memory.h>
+#include <memory/boot_allocator.h>
 #include <types.h>
 
 // ============================================================================
