+++ src/memory/internal/buddy.h
@@ -4,7 +4,7 @@
 // IMCLUDES
 // ============================================================================
 
-#include <memory.h>
+#include <memory/buddy.h>
 
 // ============================================================================
 // DEFINE AND MACRO
