+++ src/memory/page.c
@@ -1,4 +1,4 @@
-#include "page.h"
+#include "internal/page.h"
 #include "printk.h"
 
 // struct page {
